Improved support for parsing pytest test results
Added a text property to Result objects
Added a line property to TestCase objects
Those properties are set by pytest when you generate junit reports.

Added a new is_success property to TestCase objets that makes it easier
to detect if a test executed normaly.

Added a new group_by feature that is usefull when you want so split your
TestSuite objects into smaller ones. We use it to automatically generate
trackers with the result of our CI.