Add compare CLI command
`compare` creates a new report with testcases that did
not fail or error before, but do so in the newer report.

This is useful when you have a test suite that has a number
of failures that you cannot fix or skip (not ideal, but that's
sadly one of my own use cases) but you are nonetheless still
interested in regressions compared to previous runs.