junitparser.py: Fix removing items with xml.etree
The iteration in TestCase.result.setter did not work properly when using
xml.etree instead of lxml.etree. This was because the Element.remove()
method was being invoked while iterating the elements, and apparently
xml.etree does not support that.
Instead, use a prepopulated lits of Elements to iterate and remove.

Fixes #99.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>