tests: Fix test with multiple results in a TestCase
After https://github.com/weiwei/junitparser/pull/55 was merged, multiple
results are accepted in a single TestCase instance. However, this test,
which was wrong because it was invoking self.assertRaised() without a
`with` statement, remained in the tree without failing.

Instead now verify that we can import XML with multiple results in a
single TestCase instance.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>